requirements.txt
aiohttp
bs4
groq
python-dotenv
streamlit